requirements.txt
chromadb==0.4.15
voyageai==0.2.4
pandas==2.1.4
python-dotenv==1.0.0
openai==1.3.0
langchain==0.1.0
langchain-community==0.0.10
rich==13.7.0
click==8.1.7
